Move pytests for faster dev iteration
