--- py_tests/requirements.txt
@@ -1,7 +1,0 @@
-eth-utils>=1.3.0,<2
-eth-typing>=2.1.0,<3.0.0
-oyaml==0.7
-pycryptodome==3.7.3
-py_ecc>=1.6.0
-pytest>=3.6,<3.7
-../test_libs/pyspec
