+++ test_libs/pyspec/setup.py
@@ -3,6 +3,7 @@ from setuptools import setup, find_packages
 setup(
     name='pyspec',
     packages=find_packages(),
+    tests_require=["pytest"],
     install_requires=[
         "eth-utils>=1.3.0,<2",
         "eth-typing>=2.1.0,<3.0.0",
