+++ test_libs/pyspec/tests/block_processing/test_process_attestation.py
@@ -11,7 +11,7 @@ from eth2spec.phase0.spec import (
     process_attestation,
     slot_to_epoch,
 )
-from phase0.helpers import (
+from tests.helpers import (
     build_empty_block_for_next_slot,
     get_valid_attestation,
 )
